Move local_traits.py in the right place
